USB: Convert pthread_* to C++ concurrency APIs
Modernize usage of pthread APIs and convert to standard C++
facilities for threads, mutexes and condition variables.
This allows for some simiplification of mutex-protected blocks
by using scoped_lock/unique_lock.

In moving the uevent polling worker to a std::thread, an eventfd
is introduced.  This provides an alternative way to notify the
thread while it is blocked on epoll_wait() and allows elimination
of the signal handler & use of pthread_kill().

Change-Id: Ie5a113a1b1b98896e6292942993d6bc5551e4b8e