USB: Execute the HAL services as non-root
The USB HALs need not run as root and should be run at lower privilege.

In order to support this, various nodes such as ConfigFS need to be
explicitly chown'ed to the specific user/group the HAL executes under.
Within init.qcom.usb.rc, the 'mkdir' syntax to set the
{mode,owner,group} works only on the directory itself.  However, any
automatically created subdirectories that result from that must then be
'chown'ed invidually, so it does result in a more verbose init RC.

Other nodes such as sysfs should be handled from ueventd.rc.

Change-Id: I62b9647c2afa675b2db871312138fd91317f470e