init.qcom.usb.rc: Run init.qcom.usb.sh under system usb group
After the switch to non-root USB HAL, it led to the init.qcom.usb.sh script
to be unable to access the files and directories within the USB configFS.
There is a check that comes after the determination of the default USB
composition, which identifies if QDSS is present or not.  As mentioned in
the shell script, for GKI 1.0 targets, QDSS is not enabled, so ensure that
the default composition does not include it.

This results in platforms supporting 0x90DB to enumerate with 0x90B8, which
disables the QDSS interface.  This is due to the fact that the script is
unable to check for the existence of the QDSS configfs function due to
access from an invalid user (i.e. root) while configfs files are under
system usb.  Address this by modifying the exec command to execute under
the proper user and group.

Change-Id: Ic397b0bcb4870de1a3a41c93becb02d146729f25