USB: HAL: Add resetUsbPort support
In order to address Android VTS failures when running the following test
suite:
  run vts -o -m VtsAidlUsbHostTest

The USB HAL needs to properly implement the resetUsbPort callback.  The
intention of this API is to recover the USB functionality by issuing a
power on reset.  Utilize the DWC3 mode sysfs file, to switch between "none"
(disconnected) and the current operating mode.  The intended design is to
first look for the DWC3 MSM file path by using the USB controller
specified in vendor.usb.controller, and then checking if this controller
exists/binded to the msm-dwc3 platform driver.  Once identified, the USB
HAL can toggle the mode sysfs file to issue the reset.

Change-Id: I80e3b187cbdaaf1806da52eb533719c633d3e787